Rename tensor descriptor lowering pass
Rename TritonTensorDescriptorToMemref to TritonTensorDescriptorToStructured across pass registration, CMake targets, and tests.
