--- include/triton-shared/Conversion/TritonTensorDescriptorToMemref/CMakeLists.txt
@@ -1,3 +1,0 @@
-set(LLVM_TARGET_DEFINITIONS Passes.td)
-mlir_tablegen(Passes.h.inc -gen-pass-decls --name TritonTensorDescriptorToMemref)
-add_public_tablegen_target(TritonTensorDescriptorToMemrefConversionPassIncGen)
