+++ include/triton-shared/Conversion/TritonTensorDescriptorToStructured/Passes.h
@@ -5,16 +5,16 @@
 //
 //===----------------------------------------------------------------------===//
 
-#ifndef TRITON_TENSOR_DESCRIPTOR_TO_MEMREF_CONVERSION_PASSES_H
-#define TRITON_TENSOR_DESCRIPTOR_TO_MEMREF_CONVERSION_PASSES_H
+#ifndef TRITON_TENSOR_DESCRIPTOR_TO_STRUCTURED_CONVERSION_PASSES_H
+#define TRITON_TENSOR_DESCRIPTOR_TO_STRUCTURED_CONVERSION_PASSES_H
 
-#include "triton-shared/Conversion/TritonTensorDescriptorToMemref/TritonTensorDescriptorToMemref.h"
+#include "triton-shared/Conversion/TritonTensorDescriptorToStructured/TritonTensorDescriptorToStructured.h"
 
 namespace mlir {
 namespace triton {
 
 #define GEN_PASS_REGISTRATION
-#include "triton-shared/Conversion/TritonTensorDescriptorToMemref/Passes.h.inc"
+#include "triton-shared/Conversion/TritonTensorDescriptorToStructured/Passes.h.inc"
 
 } // namespace triton
 } // namespace mlir
